+++ .gitignore
@@ -12,6 +12,5 @@ docs/dist/
 docs/.DS_Store
 docs/api/.DS_Store
 .env*.local
-src/e2e-browser/testcafe-requests/
 .codesandbox/sandbox/.parcel-cache/
 test-results/
