--- .testcaferc.json
@@ -1,16 +1,0 @@
-{
-  "appCommand": "cd .codesandbox/sandbox && npx parcel --port 1234 end-to-end-test.html",
-  "appInitDelay": 50000,
-  "browsers": ["firefox:headless"],
-  "src": "src/e2e-browser/**/*.testcafe.ts",
-  "assertionTimeout": 10000,
-  "clientScripts": [
-    { "module": "@testing-library/dom/dist/@testing-library/dom.umd.js" }
-  ],
-  "compilerOptions": {
-    "typescript": {
-      "customCompilerModulePath": "../typescript"
-    }
-  },
-  "skipJsErrors": false
-}
