+++ src/e2e-browser/e2e.playwright.ts
@@ -49,7 +49,7 @@ import { config } from "dotenv-flow";
 // one 'rootDir'.
 // @ts-ignore
 import type { getHelpers } from "../../.codesandbox/sandbox/src/end-to-end-test-helpers";
-import { essUserLogin } from "./roles-playwright";
+import { essUserLogin } from "./roles";
 
 // E2eHelpers is a global defined in .codesandbox/sandbox/src/end-to-end-helper.
 // Since code inside of page.evaluate is executed in that context in the browser,
