--- src/e2e-browser/pageModels-playwright/index.ts
@@ -1,43 +1,0 @@
-/**
- * Copyright 2021 Inrupt Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- * Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-import { Page } from "@playwright/test";
-
-export class IndexPage {
-  page;
-
-  constructor(page: Page) {
-    this.page = page;
-  }
-
-  async startLogin(idp = "https://broker.pod.inrupt.com") {
-    await this.page.fill("input[type=url]", idp);
-    await this.page.click("button");
-    // The page appears stable while
-    // solid-client-authn-browser initiates the login process,
-    // so we have to explicitly wait for the redirect to the Identity Provider:
-    await this.page.waitForNavigation();
-  }
-
-  async handleRedirect() {
-    await this.page.waitForSelector("[role=alert]");
-  }
-}
