+++ src/e2e-browser/roles.ts
@@ -19,36 +19,22 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import { Role } from "testcafe";
+import { Page } from "@playwright/test";
 import { IndexPage } from "./pageModels";
 import { BrokerPage } from "./pageModels/broker";
 import { CognitoPage } from "./pageModels/cognito";
 
-export const essUser = Role("http://localhost:1234", async (t) => {
-  return await essUserLogin(t);
-});
-
-/**
- * Unfortunately solid-client-authn-browser sessions, at the time of writing,
- * do not survive a page refresh.
- * Testcafe's "Role" functionality logs you in,
- * then redirects you back to where you wanted to be for the test.
- * To avoid this redirect, the login functionality is extracted into this single helper,
- * which we'll have to use instead of the Role for now.
- * Move this function into `essUser` and replace uses of it with uses of the Role
- * once this issue is resolved: https://github.com/inrupt/solid-client-authn-js/issues/423
- */
-export const essUserLogin = async (_t: TestController) => {
-  const indexPage = new IndexPage();
+export const essUserLogin = async (page: Page) => {
+  const indexPage = new IndexPage(page);
   await indexPage.startLogin(process.env.E2E_TEST_ESS_IDP_URL);
 
-  const cognitoPage = new CognitoPage();
+  const cognitoPage = new CognitoPage(page);
   await cognitoPage.login(
     process.env.E2E_TEST_ESS_COGNITO_USER!,
     process.env.E2E_TEST_ESS_COGNITO_PASSWORD!
   );
 
-  const authorisePage = new BrokerPage();
+  const authorisePage = new BrokerPage(page);
   await authorisePage.authoriseOnce();
 
   await indexPage.handleRedirect();
